requirements.txt
pyaudio==0.2.13
colorama==0.4.6
sounddevice==0.4.6
scipy==1.10.1
pulsectl==23.5.2
python-dotenv==1.0.0
faster-whisper==0.9.0
torch==2.0.1
numpy==1.24.3
tqdm==4.66.1
soundfile==0.12.1
# Remove duplicate entries
# sounddevice (already listed above)
# scipy (already listed above)
# pulsectl (already listed above)
